pyproject.toml
[tool.poetry]
name = "debate-agent"
version = "0.1.0"
description = ""
authors = ["Richard Blythman <richardblythman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}
colorama = "^0.4.6"
python-dotenv = "^1.0.1"
litellm = "^1.44.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
